+++ lib/src/core/wo_form_cubit.dart
@@ -295,18 +295,29 @@ class WoFormValuesCubit extends Cubit<WoFormValues> {
     required String path,
     required dynamic value,
     UpdateStatus updateStatus = UpdateStatus.yes,
-  }) => onValuesChanged({path: value}, updateStatus: updateStatus);
+
+    /// If false, an update on a locked input won't have any effect.
+    bool bypassLock = false,
+  }) => onValuesChanged(
+    {path: value},
+    updateStatus: updateStatus,
+    bypassLock: bypassLock,
+  );
 
   /// **Use this method precautiously since there is no type checking !**
   void onValuesChanged(
     Json values, {
     UpdateStatus updateStatus = UpdateStatus.yes,
+
+    /// If false, an update on a locked input won't have any effect.
+    bool bypassLock = false,
   }) {
     // Remove paths of locked inputs and transform any #path
     // ignore: parameter_assignments
     values = {
       for (final entry in values.entries)
-        if (!_lockCubit.inputIsLocked(path: state.getKey(entry.key)))
+        if (bypassLock ||
+            !_lockCubit.inputIsLocked(path: state.getKey(entry.key)))
           state.getKey(entry.key): entry.value,
     };
     if (values.isEmpty) return;
@@ -468,14 +479,11 @@ class WoFormValuesCubit extends Cubit<WoFormValues> {
 
     _statusCubit._setSubmitting();
 
-    final oldLocks = _lockCubit.state;
-
-    for (final path in currentNode.getAllInputPaths(
+    final nodesLockWhileSubmitting = currentNode.getAllInputPaths(
       values: state,
       parentPath: submitPath.parentPath,
-    )) {
-      _lockCubit.lockInput(path: path);
-    }
+    );
+    _lockCubit.lockInputs(paths: nodesLockWhileSubmitting);
 
     try {
       final tempSubmitData = _tempSubmitDatas.lastOrNull;
@@ -511,11 +519,7 @@ class WoFormValuesCubit extends Cubit<WoFormValues> {
     }
 
     if (_root.uiSettings?.canModifySubmittedValues ?? true) {
-      for (final path in _lockCubit.state) {
-        if (!oldLocks.contains(path)) {
-          _lockCubit.unlockInput(path: path);
-        }
-      }
+      _lockCubit.unlockInputs(paths: nodesLockWhileSubmitting);
     }
   }
 
@@ -554,8 +558,8 @@ class WoFormValuesCubit extends Cubit<WoFormValues> {
         );
         return failure == MultistepFailure.endOfForm;
       case MultistepActionPopUntil(
-        predicate: final predicate,
-        replacementStepId: final replacementStepId,
+        :final predicate,
+        :final replacementStepId,
       ):
         final generatedSteps = state.generatedSteps;
         final currentIndex = state.currentStepIndex ?? 0;