+++ lib/src/model/data.dart
@@ -21,6 +21,7 @@ abstract class WoFormInputHeaderData with _$WoFormInputHeaderData {
     String? labelText,
     int? labelMaxLines,
     String? helperText,
+    int? helperMaxLines,
     Widget? prefixIcon,
     String? errorText,
     Widget? trailing,