+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.0.1
+
+- Added InputsNodeUiSettings.helperMaxLines
+- Hidden bottom bar when keyboard is visible
+- Added bypassLock in WoFormValuesCubit.onValueChanged
+
 ## 2.0.0
 
 - Merged wo_form_service into wo_form (DateTimeService, GooglePlaces, media handling)